Renew the LWJGL library (plus some housekeeping).
Previously, the LWJGL jars didn't include natives for MacOS. To avoid any future issues of this sort, I've redownloaded LWJGL in its complete form and placed it in the "lwjgl" folder inside "lib".

This commit also adds the build folder to the .gitignore, as well as Joe's recommendations for temporary files created by Emacs, OSX and Java.