Restructure program to use a unified main loop.
In order to implement networking, I first had to alter the program to use a main loop. That main loop calls updateGraphics(), updatePhysics() etc., and introduces the concept of a GameState. The GameState's main purpose is to hold information about the game such as positions of items and platforms.

In the process of altering the structure of the project such that it would be suitable for this style of loop, I also made many style changes including renaming methods, adding JavaDocs and breaking down large methods into smaller utility methods.

This commit is essentially a marker for where the game runs as it did before the changes. (I actually think it runs more smoothly, although that might be my imagination!)