Fixed the main problem with networked objects.
The problem we were facing was that despite the server updating the game state, whenever that state was sent to the client, it appeared as the same initial game state.

For a long time, I thought this was some problem with serialization; maybe some default behaviour when values couldn't be serialized so a constructor was called instead. But this didn't make a huge amount of sense, because the value that kept being sent was (400, 400) for the ball, which didn't appear in any constructor.

As it turns out Java does some caching behind the scenes of ObjectOutputStream instances. So when I sent my GameState object, the ObjectOutputStream would retrieve the GameState from a previous frame from its cache and just send an object reference, instead of the full updated GameState.

This problem was solved by adding a call to `outputStream.reset()` after every call of `outputStream.writeObject(o)`. This invalidates the cache after every object-write, and means that the full data will be sent each time.