Remove regular platform randomness and fix timestep.
Previously, platforms would update their position based on whether they were off-screen, generating random X-coordinates for themselves. This clearly can't work when both the client and server must be running Platform.update().

This commit changes things so that the platforms are randomly generated at the start, then repeat in a loop. This is less interesting for sure, but it works for a demo. In future if we want more interesting patterns of platform, we can generate them at the start.

This commit also fixes the timestep to something playable. Still outstanding is the task to fix the timestep given variable time taken, using the last frame's render time to affect the physics simulation etc.